requirements.txt
absl-py==0.11.0
albumentations==1.0.3
argon2-cffi==21.1.0
astunparse==1.6.3
attrs==20.3.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
clang==5.0
colorama==0.4.4
cycler==0.10.0
Cython==0.29.24
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.3
easyocr==1.5.0
entrypoints==0.3
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.40.0
h5py==3.1.0
idna==2.10
imageio==2.9.0
importlib-resources==5.1.2
imutils==0.5.4
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
keras-flops==0.1.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.8.4
notebook==6.4.6
numexpr==2.8.3
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.1.1
pip==22.2
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.22
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycparser==2.21
pydot==1.4.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.3
PyWavelets==1.3.0
pywin32==302
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.4.1
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==63.2.0
six==1.15.0
sklearn==0.0
tables==3.7.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-addons==0.14.0
tensorflow-datasets==4.2.0
tensorflow-estimator==2.6.0
tensorflow-metadata==0.30.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tf-slim==1.1.0
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.12.0
torchvision==0.13.0
tornado==6.1
tqdm==4.60.0
traitlets==5.1.1
typeguard==2.12.0
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.36.2
widgetsnbextension==3.5.2
wrapt==1.12.1
